requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
autobahn==25.10.2
Automat==25.4.16
billiard==4.2.2
cachetools==6.2.1
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.2.1
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
Django==5.2.8
django-celery-beat==2.8.1
django-debug-toolbar==6.1.0
django-extensions==4.1
django-redis==6.0.0
django-timezone-field==7.1
django_celery_results==2.6.0
dnspython==2.8.0
executing==2.2.1
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-google-genai==3.0.3
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langchain-weaviate==0.0.6
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
marshmallow==3.26.1
matplotlib-inline==0.2.1
motor==3.7.1
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.8.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.4
PyMuPDF==1.26.6
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
txaio==25.9.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
vine==5.1.0
wcwidth==0.2.14
weaviate-client==4.18.0
whitenoise==6.11.0
xxhash==3.6.0
yarl==1.22.0
zope.interface==8.1
zstandard==0.25.0
